+++ src/main/java/ProjetoMetci/metci/NodePower.java
@@ -1,37 +1,37 @@
 package ProjetoMetci.metci;
 
 public class NodePower {
-	private long processor;
-	private long ram;
-	private long hd;
+	private double processor;
+	private double ram;
+	private double hd;
 
-	public NodePower(long processor, long ram, long hd) {
+	public NodePower(double processor, double ram, double hd) {
 		this.processor = processor;
 		this.ram = ram;
 		this.hd = hd;
 	}
 
-	public long getProcessor() {
+	public double getProcessor() {
 		return processor;
 	}
 
-	public void setProcessor(long processor) {
+	public void setProcessor(double processor) {
 		this.processor = processor;
 	}
 
-	public long getRam() {
+	public double getRam() {
 		return ram;
 	}
 
-	public void setRam(long ram) {
+	public void setRam(double ram) {
 		this.ram = ram;
 	}
 
-	public long getHd() {
+	public double getHd() {
 		return hd;
 	}
 
-	public void setHd(long hd) {
+	public void setHd(double hd) {
 		this.hd = hd;
 	}
